requirements.txt
openai>=1.12.0
moviepy>=1.0.3
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
schedule>=1.2.0
mutagen>=1.47.0
